+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/Helpers.scala
@@ -6,6 +6,160 @@ import scalaz._
 import Scalaz._
 
 object Helpers {
+  val MaxExpandRounds: Int = 10
+
+  def maybeExpandInt(
+      current: Option[Int],
+      limit: Option[Int],
+      exclusive: Boolean,
+      round: Int = 1
+  ): (Option[Int], Boolean) = {
+    if (round > MaxExpandRounds) {
+      (None, false)
+    } else {
+      (current, limit) match {
+        case (Some(currentInt), Some(otherInt)) =>
+          expandInt(currentInt, otherInt, exclusive, round)
+
+        // We have no limit, keep it that way
+        case (None, _) => (None, false)
+
+        // There is no limit on the other schema, so we can't have one here
+        case (_, None) => (None, false)
+      }
+    }
+  }
+
+  def expandInt(
+      current: Int,
+      limit: Int,
+      exclusive: Boolean,
+      round: Int
+  ): (Option[Int], Boolean) = {
+    if (current < 0) {
+      // TODO We can probably make better choices for negative values
+      maybeExpandInt(Some(0), Some(limit), exclusive, round + 1)
+    } else if (exclusive && current < limit) {
+      maybeExpandInt(Some(current), Some(limit), false, round + 1)
+    } else if (
+      (exclusive && current > limit) || (!exclusive && current >= limit)
+    ) {
+      (Some(current), exclusive)
+    } else if (current < 10) {
+      maybeExpandInt(Some(current + 1), Some(limit), exclusive, round + 1)
+    } else {
+      // Since 2 is a common power, consider the next closest power
+      // but first consider one less since this occurs frequently
+      // (e.g. 255, 65535)
+      val closestPow2 =
+        Math.pow(2, (Math.log(current) / Math.log(2)).ceil).toInt
+      val next2 = if ((closestPow2 - 1) > current) {
+        closestPow2 - 1
+      } else {
+        closestPow2
+      }
+
+      // Otherwise, increase by the closest power of 10 (examples below)
+      // 12 => 13
+      // 20 => 30
+      // 99 => 100
+      // 110 => 200
+      // 217 => 300
+      val pow10 = Math.pow(10, Math.log10(current).floor - 1).toInt
+      val next10 = current + pow10 - (current % pow10)
+      if (next2 < next10 && next2 > current) {
+        maybeExpandInt(Some(next2), Some(limit), exclusive, round + 1)
+      } else {
+        maybeExpandInt(Some(next10), Some(limit), exclusive, round + 1)
+      }
+    }
+  }
+
+  def maybeContractInt(
+      current: Option[Int],
+      limit: Option[Int],
+      exclusive: Boolean,
+      round: Int = 1
+  ): (Option[Int], Boolean) = {
+    if (round > MaxExpandRounds) {
+      (None, false)
+    } else {
+      (current, limit) match {
+        case (Some(currentInt), Some(otherInt)) =>
+          contractInt(currentInt, otherInt, exclusive, round)
+
+        // We have no limit, keep it that way
+        case (None, _) => (None, false)
+
+        // There is no limit on the other schema, so we can't have one here
+        case (_, None) => (None, false)
+      }
+    }
+  }
+
+  @SuppressWarnings(Array("org.wartremover.warts.Equals"))
+  def contractInt(
+      current: Int,
+      limit: Int,
+      exclusive: Boolean,
+      round: Int
+  ): (Option[Int], Boolean) = {
+    if (current < 0) {
+      // TODO We can probably make better choices for negative values
+      maybeContractInt(Some(0), Some(limit), exclusive, round + 1)
+    } else if (exclusive && current < limit) {
+      maybeContractInt(Some(current), Some(limit), false, round + 1)
+    } else if (
+      (exclusive && current < limit) || (!exclusive && current <= limit)
+    ) {
+      (Some(current), exclusive)
+    } else if (current < 10) {
+      maybeContractInt(Some(current - 1), Some(limit), exclusive, round + 1)
+    } else {
+      // Since 2 is a common power, consider the next smallest power
+      // but first consider just one less since this occurs frequently
+      // (e.g. 255, 65535)
+      val log2 = Math.log(current) / Math.log(2)
+      val closestPow2 =
+        Math.pow(2, (log2).ceil - 1).toInt
+      val next2 = if (log2.isValidInt) {
+        current - 1
+      } else {
+        closestPow2
+      }
+
+      // Otherwise, decrease by the closest power of 10 (examples below)
+      // 13 => 12
+      // 30 => 20
+      // 100 => 99
+      // 200 => 190
+      // 307 => 300
+      val pow10 = Math.pow(10, Math.log10(current).floor - 1).toInt
+      val next10 = if (current % pow10 == 0) {
+        current - pow10
+      } else {
+        current - (current % pow10)
+      }
+      if (next2 > next10) {
+        maybeContractInt(Some(next2), Some(limit), exclusive, round + 1)
+      } else {
+        maybeContractInt(Some(next10), Some(limit), exclusive, round + 1)
+      }
+    }
+  }
+
+  // Source: https://stackoverflow.com/a/30281343/123695
+  @SuppressWarnings(Array("org.wartremover.warts.Recursion"))
+  def factorize(x: Int): List[Int] = {
+    def foo(x: Int, a: Int): List[Int] = if (a * a > x) List(x)
+    else
+      x % a match {
+        case 0 => a :: foo(x / a, a)
+        case _ => foo(x, a + 1)
+      }
+    foo(x, 2).sorted
+  }
+
   @SuppressWarnings(Array("org.wartremover.warts.OptionPartial"))
   def isMinCompatibleWith[A: Order](
       value1: Option[A],