+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/StringSchemaSpec.scala
@@ -92,6 +92,18 @@ class StringSchemaSpec extends UnitSpec {
     (formatProp.toJson \ "format") shouldBe JNothing
   }
 
+  it should "not expand for compatible formats" in {
+    FormatProperty(Map("url" -> 100))
+      .expandTo(FormatProperty(Map("url" -> 10)))
+      .maxFormat() shouldBe Some("url")
+  }
+
+  it should "expand to remove formats if needed" in {
+    FormatProperty(Map("url" -> 100))
+      .expandTo(FormatProperty(Map("email" -> 10)))
+      .maxFormat() shouldBe None
+  }
+
   behavior of "StaticPatternProperty"
 
   it should "should not allow merging regex properties" in {
@@ -128,6 +140,13 @@ class StringSchemaSpec extends UnitSpec {
     for (_ <- 1 to 10) { prefixSchema = prefixSchema.mergeValue("bazfoo") }
     val prefixProp = prefixSchema.get[PatternProperty]
     (prefixProp.toJson \ "pattern").extract[String] shouldBe "^ba.*foo$"
+  }
+
+  it should "expand to remove patterns if needed" in {
+    val newPattern =
+      PatternProperty(Some("foo"), Some("bar")).expandTo(PatternProperty())
+    newPattern.prefix shouldBe None
+    newPattern.suffix shouldBe None
   }
 
   behavior of "StringLengthHistogramProperty"