+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow the maximum number of examples to be configured
 - Allow configuration of `additionalProperties`
 - Allow checking of schema compatibility
+- Expand schema properties to cover another schema where possible
 
 ### Changed
 - Use ranking to improve possible primary key suggestions